Version 4.1.2 and fix build errors
