+++ app/config/config.json
@@ -1,6 +1,6 @@
 {
-  "IS_LOCAL": true,
+  "IS_LOCAL": false,
   "IS_PROD": false,
   "IS_CANARY": false,
-  "BUILD_VERSION": "4.1.0"
+  "BUILD_VERSION": "4.1.2"
 }
