+++ app/containers/MassEnergizeSuperAdmin/ME  Tools/media library/shared/components/library/Library.js
@@ -6,7 +6,7 @@ import { blank } from "../../utils/values";
 import {
   LoadMoreContainer,
   ProgressCircleWithLabel,
-} from "../../../../../Gallery/Gallery";
+} from "../../../../../Gallery/utils";
 
 function Library({
   multiple,
