+++ deployment/build.json
@@ -1,5 +1,5 @@
 {
-  "dev": "4.1.0",
+  "dev": "4.1.2",
   "prod": "4.0.0",
-  "canary": "4.1.0"
+  "canary": "4.1.2"
 }
