+++ package.json
@@ -88,19 +88,19 @@
     ]
   },
   "dependencies": {
-    "@babel/plugin-proposal-export-default-from": "^7.12.1",
-    "@babel/plugin-proposal-object-rest-spread": "^7.12.1",
+    "@babel/plugin-proposal-export-default-from": "^7.16.5",
+    "@babel/plugin-proposal-object-rest-spread": "^7.16.5",
     "@babel/polyfill": "7.4.3",
     "@date-io/moment": "^1.3.13",
-    "@material-ui/core": "^3.0.0",
+    "@material-ui/core": "^3.9.4",
     "@material-ui/icons": "^3.0.2",
     "@material-ui/lab": "^3.0.0-alpha.28",
-    "@tinymce/tinymce-react": "^3.8.1",
-    "@types/googlemaps": "^3.40.3",
+    "@tinymce/tinymce-react": "^3.13.0",
+    "@types/googlemaps": "^3.43.3",
     "@types/markerclustererplus": "^2.1.33",
     "acorn": "^6.4.2",
-    "autosuggest-highlight": "^3.1.1",
-    "axios": "^0.21.1",
+    "autosuggest-highlight": "^3.2.0",
+    "axios": "^0.21.4",
     "babel-polyfill": "6.26.0",
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
@@ -108,7 +108,7 @@
     "bourbon-neat": "^3.0.1",
     "chalk": "^2.4.2",
     "chart.js": "^2.9.4",
-    "classnames": "^2.2.6",
+    "classnames": "^2.3.1",
     "compression": "1.7.4",
     "connected-react-router": "6.4.0",
     "cross-env": "5.2.0",
@@ -134,16 +134,16 @@
     "ip": "1.1.5",
     "js-cookie": "^2.2.1",
     "jss": "^9.8.7",
-    "keycode": "^2.2.0",
+    "keycode": "^2.2.1",
     "lodash": "^4.17.21",
     "material-ui": "^0.20.2",
     "material-ui-pickers": "^2.2.4",
     "minimist": "1.2.3",
     "moment": "^2.24.0",
-    "moment-timezone": "^0.5.31",
+    "moment-timezone": "^0.5.34",
     "mui-datatables": "^2.15.0",
     "prop-types": "15.7.2",
-    "qs": "^6.9.4",
+    "qs": "^6.10.2",
     "rcolor": "^1.0.1",
     "react": "16.8.6",
     "react-anchor-link-smooth-scroll": "^1.0.12",
@@ -151,27 +151,27 @@
     "react-autosuggest": "^9.4.3",
     "react-big-calendar": "^0.19.1",
     "react-calendar": "^2.19.2",
-    "react-chartjs-2": "^2.10.0",
+    "react-chartjs-2": "^2.11.2",
     "react-charts": "^1.0.10",
     "react-clock": "^2.4.0",
     "react-countup": "^3.0.3",
     "react-dom": "16.8.6",
-    "react-draft-wysiwyg": "^1.14.6",
+    "react-draft-wysiwyg": "^1.14.7",
     "react-dropzone": "^4.3.0",
     "react-event-listener": "^0.6.6",
     "react-google-maps": "^9.4.5",
     "react-helmet": "6.0.0-beta",
-    "react-image-crop": "^8.6.6",
+    "react-image-crop": "^8.6.12",
     "react-intl": "2.8.0",
     "react-ionicons": "^2.1.6",
     "react-loading-bar": "0.0.7",
     "react-markdown": "^3.1.5",
-    "react-modal": "^3.11.2",
+    "react-modal": "^3.14.4",
     "react-number-format": "^3.6.2",
-    "react-papaparse": "^3.16.1",
+    "react-papaparse": "^3.18.1",
     "react-popper": "^0.10.4",
     "react-redux": "7.0.2",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^5.3.0",
     "react-scroll-parallax": "^1.3.5",
     "react-scrollspy": "^3.4.3",
     "react-select": "^1.2.1",
@@ -188,7 +188,7 @@
     "redux-form-material-ui": "^5.0.0-beta.3",
     "redux-immutable": "4.0.0",
     "redux-saga": "1.0.2",
-    "redux-thunk": "^2.3.0",
+    "redux-thunk": "^2.4.1",
     "reselect": "4.0.0",
     "sanitize.css": "8.0.0",
     "serve-favicon": "^2.4.5",
@@ -204,7 +204,7 @@
     "@babel/plugin-transform-modules-commonjs": "7.4.3",
     "@babel/plugin-transform-react-constant-elements": "7.2.0",
     "@babel/plugin-transform-react-inline-elements": "7.2.0",
-    "@babel/preset-env": "^7.16.4",
+    "@babel/preset-env": "^7.16.5",
     "@babel/preset-react": "7.0.0",
     "@babel/register": "7.4.0",
     "add-asset-html-webpack-plugin": "3.1.3",
@@ -256,17 +256,17 @@
     "postcss-loader": "^2.1.6",
     "pre-commit": "1.2.2",
     "prettier": "1.17.0",
-    "prismjs": "^1.22.0",
+    "prismjs": "^1.25.0",
     "raw-loader": "^0.5.1",
     "react-app-polyfill": "0.2.2",
     "react-test-renderer": "16.8.6",
     "react-testing-library": "6.1.2",
-    "react-to-print": "^2.11.0",
+    "react-to-print": "^2.14.2",
     "rimraf": "2.6.3",
     "sass-loader": "^7.3.1",
     "sass-material-colors": "0.0.5",
     "shelljs": "^0.8.4",
-    "source-map-loader": "^1.0.0",
+    "source-map-loader": "^1.1.3",
     "style-loader": "0.23.1",
     "stylelint": "10.0.1",
     "stylelint-config-recommended": "2.2.0",
@@ -279,7 +279,7 @@
     "webpack-cli": "^3.3.12",
     "webpack-dev-middleware": "3.6.2",
     "webpack-hot-middleware": "2.24.3",
-    "webpack-pwa-manifest": "^4.2.0",
+    "webpack-pwa-manifest": "^4.3.0",
     "whatwg-fetch": "3.0.0"
   }
 }
